Add BeefyZapSolidlyDyst.sol (#224)
* Create BeefyZapSolidlyDyst.sol

Noticed a deployed and active zap contract variant that wasn't in the main repo, so extracted a copy from the block explorer (see https://polygonscan.com/address/0x1b03127ec0f4d12efd84d84c70d5aa7ef83fe136#code). Please note that this is a contract deployed by Wivern, and not by me. As it has been live for some time and conducted thousands of transactions, I thought it would be useful to add to the main repo.

* Change path for imports and rename file

---------

Co-authored-by: kexley <sircryptomaxi@gmail.com>